s3_feat_110_Edit-infor-teamdetail_bon

+++ service/teamDetail.ts
@@ -2,13 +2,10 @@ import api from "@/config/api";
 import { KickMemberType } from "@/types";
 
 const TeamDetailApi = {
-  kickMember: (memberId: number, kickData: KickMemberType) =>
-    api.delete(`/team-member/${memberId}/kick`, { data: kickData }),
-
-  deleteTeam: (teamId: number, userId: number) =>
-    api.delete(`/team/${teamId}?userId=${userId}`),
-  
+  kickMember: (memberId: number, kickData: KickMemberType) => api.delete(`/team-member/${memberId}/kick`, { data: kickData }),
+  deleteTeam: (teamId: number, userId: number) => api.delete(`/team/${teamId}?userId=${userId}`),
   getTeamDetail: (teamId: number, userId: number ) => api.get(`/team/${teamId}/user/${userId}`),
+  updateTeam: (teamId: number, formData: object) => api.patch(`/team/${teamId}`, formData),
 };
 
 export default TeamDetailApi;
